# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ensure API
  version: 2.0.0
  description: >
    Backend API for the note web app (foundation for a future deadman switch).
    Each signed-in user has at most one note of their own, stored encrypted at rest
    (feature 004-note-ownership-encryption); reads and writes are scoped to the
    authenticated user, so no endpoint can address another user's note. Gated behind
    Google SSO (feature 002-google-sso-auth): sign-in is Google-only; the access token
    is a ~1h JWT carried in an httpOnly cookie, silently refreshed while the user is
    active, with a ~24h inactivity window.
servers:
  - url: /api
    description: API root (same-origin in dev and single-instance deploy)
paths:
...
  /contact:
    get:
      operationId: getContacts
      summary: Get the caller's own contact list
      description: >
        Returns the authenticated caller's own contacts (may be empty), ordered by
        creation time. Contacts are always scoped to the caller — another user's
        contacts are never returned. Requires a valid access token.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The caller's contacts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
              examples:
                some:
                  summary: A few email contacts
                  value:
                    contacts:
                      - id: "f3b1c2d4-0000-4000-8000-000000000001"
                        type: "email"
                        value: "Alice@Example.com"
                        createdAt: "2026-06-07T09:12:00.000Z"
                empty:
                  summary: No contacts yet
                  value:
                    contacts: []
    post:
      operationId: addContact
      summary: Add a contact to the caller's own list
      description: >
        Adds a contact (only type "email" is accepted in this release) to the caller's
        list and returns the stored contact. The email is stored as entered (trimmed,
        original case preserved); duplicate detection is case-insensitive. Rejects an
        unsupported type, a malformed email, a value longer than 320 characters, a
        duplicate, or an add that would exceed the 50-contact limit. Requires a valid
        access token.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
            examples:
              valid:
                value:
                  type: "email"
                  value: "alice@example.com"
      responses:
        "201":
          description: The contact was added; returns the stored contact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
              example:
                id: "f3b1c2d4-0000-4000-8000-000000000001"
                type: "email"
                value: "alice@example.com"
                createdAt: "2026-06-07T09:12:00.000Z"
        "400":
          description: Validation failed (missing fields, unsupported type, malformed email, or too long).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalidEmail:
                  value:
                    error: "VALIDATION_ERROR"
                    message: "Enter a valid email address."
                tooLong:
                  value:
                    error: "VALIDATION_ERROR"
                    message: "Email must be at most 320 characters."
        "409":
          description: The contact already exists, or the caller has reached the 50-contact limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                duplicate:
                  value:
                    error: "DUPLICATE_CONTACT"
                    message: "That contact is already in your list."
                limit:
                  value:
                    error: "CONTACT_LIMIT_REACHED"
                    message: "You can have at most 50 contacts."
        "401":
          description: No valid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "UNAUTHORIZED", message: "Sign in to continue." }
  /contact/{id}:
    delete:
      operationId: removeContact
      summary: Remove a contact from the caller's own list
      description: >
        Removes the identified contact from the caller's list. Idempotent: returns 204
        whether or not a matching contact existed. A contact owned by another user is
        treated exactly like a non-existent one (204 no-op), so one user can never
        affect or probe another's contacts. Requires a valid access token.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          description: The id of the contact to remove.
      responses:
        "204":
          description: The contact was removed, or no matching contact existed (idempotent).
...
  /contact/{id}/verify:
    post:
      operationId: sendContactVerification
      summary: Send (or resend) a verification email to one of the caller's contacts
      description: >
        Mints a high-entropy token, stores only its SHA-256 hash plus a future expiry on the
        owned contact, and sends a verification email through the generic notify() dispatcher
        (email channel) whose body carries a single APP_BASE_URL link with the raw token shown
        once. Scoped to the caller (`req.user.id`): a contact id that is not the caller's own
        is treated as not-found (404) and no email is sent. Resending is allowed and refreshes
        the token (overwriting the prior hash + expiry), invalidating any earlier link. Requires
        a valid access token.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          description: The id of the caller's contact to verify.
      responses:
        "200":
          description: The verification email was dispatched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactVerifySendResponse"
              example: { sent: true }
        "401":
          description: No valid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "UNAUTHORIZED", message: "Sign in to continue." }
        "404":
          description: No such contact for the caller (no email sent).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "NOT_FOUND", message: "Contact not found." }
        "502":
          description: The verification email could not be dispatched; the contact stays unverified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: { error: "EMAIL_SEND_FAILED", message: "Could not send the verification email. Please try again." }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: >
        httpOnly, Secure, SameSite=Lax cookie holding the ~1h access-token JWT. Set by the Google
        callback (and by the test-login route in test mode), refreshed via POST /auth/refresh.
  schemas:
...
    Contact:
      type: object
      additionalProperties: false
      required: [id, type, value, createdAt, verified, verifiedAt]
      properties:
        id:
          type: string
          description: Server-assigned unique id for this contact (used to remove it).
        type:
          type: string
          enum: [email]
          description: >
            Contact type. Only "email" is supported in this release; the field exists so
            future types (phone, username) can be added without changing existing contacts.
        value:
          type: string
          maxLength: 320
          description: The email address as entered (trimmed, original case preserved for display).
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp of when the contact was added.
        verified:
          type: boolean
          description: >
            Whether the contact has proven control of its address (feature 009). Derived
            from `verified_at != null`; every pre-existing contact is unverified by default.
            Only verified contacts ever receive a release (feature 010).
        verifiedAt:
          type: string
          format: date-time
          nullable: true
          description: >
            ISO 8601 UTC timestamp of when the contact was verified, or null if unverified.
            The internal verification token hash and expiry are never serialized.
    ContactInput:
      type: object
      additionalProperties: false
      required: [type, value]
      properties:
        type:
          type: string
          enum: [email]
          description: Only "email" is accepted in this release.
        value:
          type: string
          minLength: 1
          maxLength: 320
          description: A well-formed email address.
    ContactListResponse:
      type: object
      additionalProperties: false
      required: [contacts]
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
          description: The caller's contacts, ordered by creation time (may be empty).
    ContactVerifySendResponse:
      type: object
      additionalProperties: false
      required: [sent]
      properties:
        sent:
          type: boolean
          description: >
            True when the verification email was dispatched through the generic notify()
            dispatcher to the contact's address (feature 009).
...
    Error:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
          examples: ["UNAUTHORIZED", "VALIDATION_ERROR", "CHANNEL_NOT_SUPPORTED"]
        message:
          type: string
          description: Human-readable explanation suitable for display.
...